openingd: pull out common code, rename some things
We're going to reuse all this code for dualopend, which is coming soon.
