+++ lightningd/Makefile
@@ -24,6 +24,7 @@ LIGHTNINGD_SRC :=				\
 	lightningd/memdump.c			\
 	lightningd/notification.c		\
 	lightningd/onchain_control.c		\
+	lightningd/opening_common.c		\
 	lightningd/opening_control.c		\
 	lightningd/options.c			\
 	lightningd/pay.c			\
