+++ lightningd/peer_control.c
@@ -49,6 +49,7 @@
 #include <lightningd/memdump.h>
 #include <lightningd/notification.h>
 #include <lightningd/onchain_control.h>
+#include <lightningd/opening_common.h>
 #include <lightningd/opening_control.h>
 #include <lightningd/options.h>
 #include <lightningd/peer_htlcs.h>
